refactor: split endCombatPhaseCommand into focused modules
Extract combat end handling and phase transition logic into separate modules:

- combatEndHandlers.ts: Combat end resolution, conquest, withdrawal, monastery burning
- phaseTransitions.ts: Phase state machine and transition handlers

Consolidates duplicate fame/reputation pattern into applyDefeatedEnemyRewards().

Reduces endCombatPhaseCommand.ts from 599 lines to 52 lines while maintaining
the same public API.
